move StylizedBaseDialog to sharedComponents
