+++ apps/src/lib/levelbuilder/code-docs-editor/ProgrammingEnvironmentsTable.jsx
@@ -2,8 +2,8 @@ import PropTypes from 'prop-types';
 import React, {useState} from 'react';
 import * as Table from 'reactabular-table';
 
-import StylizedBaseDialog from '@cdo/apps/componentLibrary/StylizedBaseDialog';
 import Button from '@cdo/apps/legacySharedComponents/Button';
+import StylizedBaseDialog from '@cdo/apps/sharedComponents/StylizedBaseDialog';
 import FontAwesome from '@cdo/apps/templates/FontAwesome';
 
 const destroyEnvironment = (destroyPath, successCallback, failureCallback) => {
