+++ apps/src/lib/levelbuilder/code-docs-editor/ProgrammingExpressionsTable.jsx
@@ -3,8 +3,8 @@ import queryString from 'query-string';
 import React, {useEffect, useState} from 'react';
 import * as Table from 'reactabular-table';
 
-import StylizedBaseDialog from '@cdo/apps/componentLibrary/StylizedBaseDialog';
 import Button from '@cdo/apps/legacySharedComponents/Button';
+import StylizedBaseDialog from '@cdo/apps/sharedComponents/StylizedBaseDialog';
 import PaginationWrapper from '@cdo/apps/templates/PaginationWrapper';
 
 import CloneProgrammingExpressionDialog from './CloneProgrammingExpressionDialog';
