+++ apps/src/p5lab/AnimationPicker/AnimationPicker.jsx
@@ -3,8 +3,8 @@ import React from 'react';
 import {connect} from 'react-redux';
 
 import HiddenUploader from '@cdo/apps/code-studio/components/HiddenUploader';
-import StylizedBaseDialog from '@cdo/apps/componentLibrary/StylizedBaseDialog';
 import {AnimationProps} from '@cdo/apps/p5lab/shapes';
+import StylizedBaseDialog from '@cdo/apps/sharedComponents/StylizedBaseDialog';
 import BaseDialog from '@cdo/apps/templates/BaseDialog.jsx';
 import {createUuid, makeEnum} from '@cdo/apps/utils';
