+++ apps/src/sharedComponents/StylizedBaseDialog.jsx
