+++ apps/src/sharedComponents/StylizedBaseDialog.story.jsx
